requirements.txt
aiohttp==3.7.4.post0
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
idna==3.3
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.2.0
nextcord==2.0.0a3
py-cord==1.7.3
pylint==2.6.0
python-dotenv==0.19.2
six==1.15.0
toml==0.10.2
typing_extensions==4.0.0
wrapt==1.12.1
yarl==1.7.2
